requirements.txt
streamlit>=1.42.0
pandas>=2.2.0
langchain>=0.3.20
langchain-openai>=0.3.7
langchain-groq>=0.2.4
langchain-google-genai>=2.1.10
python-dotenv>=1.0.1
pymongo>=4.10.1
